Add alert dialog and add it to announcement deletion

+++ frontend/src/app/(dashboard)/home/page.tsx
@@ -24,6 +24,17 @@ import { isFacultyOrAdmin } from "@/lib/rbac";
 import { Trash2 } from "lucide-react";
 import { useFetchCollection } from "@/lib/hooks";
 import GenericError from "@/components/GenericError";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogTrigger,
+  AlertDialogTitle,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogDescription,
+  AlertDialogCancel,
+  AlertDialogContent,
+} from "@/components/ui/alert-dialog";
 
 export default function Home() {
   const { user, token } = useAuth();
@@ -113,14 +124,30 @@ export default function Home() {
               </TableCell>
               <TableCell className="flex justify-end h-full items-center gap-2">
                 {isAuthorized && (
-                  <Button
-                    // TODO: Show warning dialog
-                    onClick={() => deleteAnnouncementById(element.id)}
-                    variant={"destructive"}
-                    size={"icon"}
-                  >
-                    <Trash2 />
-                  </Button>
+                  <AlertDialog>
+                    <AlertDialogTrigger asChild>
+                      <Button variant={"destructive"} size={"icon"}>
+                        <Trash2 />
+                      </Button>
+                    </AlertDialogTrigger>
+                    <AlertDialogContent>
+                      <AlertDialogHeader>
+                        <AlertDialogTitle>Are you sure?</AlertDialogTitle>
+                        <AlertDialogDescription>
+                          This action cannot be undone.
+                        </AlertDialogDescription>
+                      </AlertDialogHeader>
+                      <AlertDialogFooter>
+                        <AlertDialogCancel>Cancel</AlertDialogCancel>
+                        <AlertDialogAction
+                          className="bg-destructive hover:bg-destructive/90 text-destructive-foreground"
+                          onClick={() => deleteAnnouncementById(element.id)}
+                        >
+                          Continue
+                        </AlertDialogAction>
+                      </AlertDialogFooter>
+                    </AlertDialogContent>
+                  </AlertDialog>
                 )}
                 {/* FIXME: Sheet open & openChanged cause stale content, make a list of open instead */}
                 <Sheet>
